requirements.txt
ffmpeg-python
numpy
pandas
torch
pyannote.audio
transformers
accelerate